+++ demo-setup/start-demo-with-meddream.sh
@@ -3,7 +3,7 @@ set -o xtrace
 set -o errexit
 
 pushd ../sources
-docker build -t orthancteam/orthanc-nginx -f Dockerfile.orthanc-nginx .
+docker build -t orthancteam/orthanc-nginx -f nginx/Dockerfile.orthanc-nginx .
 docker build -t orthancteam/orthanc-token-service -f Dockerfile.orthanc-token-service .
 docker build -t orthancteam/orthanc-anonymizer  -f Dockerfile.orthanc-anonymizer .
 docker build -t orthancteam/orthanc-share-landing  -f Dockerfile.orthanc-share-landing .